Refactor/hdf5tableio (#671)
* separate TableWriter and TableReader base classes into separate file (tableio.py)
* rename hdftableio.py -> hdf5tableio.py
*  fix some imports to be relative
*  fix some style warnings
*  change log statments to not use .format
