+++ ctapipe/io/__init__.py
@@ -1,14 +1,18 @@
-from .eventsource import EventSource
-from .hessioeventsource import HESSIOEventSource
 from .array import get_array_layout
-from .eventsourcefactory import EventSourceFactory, event_source
 from .eventseeker import EventSeeker
-from .hdftableio import (
-    TableReader, TableWriter, HDF5TableReader, HDF5TableWriter
-)
-
+from .eventsource import EventSource
+from .eventsourcefactory import EventSourceFactory, event_source
+from .hdf5tableio import HDF5TableReader, HDF5TableWriter
+from .hessioeventsource import HESSIOEventSource
+from .tableio import TableWriter, TableReader
 
 __all__ = [
-    'TableWriter', 'TableReader', 'HDF5TableWriter', 'HDF5TableReader',
-    'EventSeeker', 'EventSourceFactory', 'EventSource', 'event_source'
+    'HDF5TableWriter',
+    'HDF5TableReader',
+    'TableWriter',
+    'TableReader',
+    'EventSeeker',
+    'EventSourceFactory',
+    'EventSource',
+    'event_source'
 ]
